+++ CHANGELOG.md
@@ -9,11 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Flake now exports NixOS module
+
+### Changed
 - support defining a postgres password in config or env
 
 ### Fixed
 
-- bump blockfrost-utils to 2.6.2
+- nixpkgs updated to latest `nixos-23.05`
+- NodeJS updated to `18.16.1`
+- blockfrost-utils updated to `2.6.2`
 
 ## [1.6.0] - 2023-06-28